DTO de productos, ítems, Carrito, checkout

+++ back-ecommerce/src/main/java/com/uade/ecommerce/controller/CarritoController.java
@@ -10,6 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/carritos")
 public class CarritoController {
@@ -82,7 +84,7 @@ public class CarritoController {
     public ResponseEntity<CheckoutRespuestaDTO> checkout(
             @PathVariable Long usuarioId
     ) {
-        Double total =
+        BigDecimal total =
                 carritoService.checkout(usuarioId);
 
         CheckoutRespuestaDTO respuesta =
